backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-decouple==3.8
aiohttp==3.9.1
google-generativeai==0.8.3
gigachat==0.1.34